* Rewrite ipsec.conf.common to work around connaddrfamily issues
connaddrfamily= gets confusing via many also= loads because in some
cases it cannot and in some cases it should be there. It should really
be phased out. Once we know left/leftsubnet/right/rightsubnet, we
can deduce the connaddrfamily. Though a little more complicated on %any
and %defaultroute conns.

Conflicts:

	testing/baseconfigs/all/etc/ipsec.d/ipsec.conf.common