* Attempt to flush the unbound cache for DNS domains obtained via XAUTH
When we receive a domain via PLUTO_CISCO_DOMAIN_INFO, in the updown
scripts we now check if there is a local unbound resolver running. If so,
we flush the cache of that domain (and every subdomain of it).

We do the same on disconnect of teh VPN.

This will ensure that unbound does not have internal IP entries in its
cache when dropping the VPN.

Note that we are waiting on unbound to finish implementing the method:

unbound-control forward_zone redhat.com 1.2.3.4 5.6.7.8

That way, once we connect the VPN, we can dynamically reconfigure unbound
to forward all queries for the PLUTO_CISCO_DOMAIN_INFO to the servers
we received in $PLUTO_CISCO_DNS_INFO.

This will even work when using multiple tunnels to different domains!

A similar updown script should be written for pppd to ship with xl2tpd,
so it does the same for IPsec/L2TP connections.