* IKEv2: PARENT_R1 state inserted connection hash based on zero rcookie
This caused the 2nd packet to ignore the first state, as the lookup
was done with the icookie/rcookie. Since calculating an rcookie is
relatively cheap (and we have the 6msg exchange when under DDOS attack)
we now generate the rcookie immediate before initialize_new_state()

This meant removing a bogus call to rehash_state() later on.

It also copied the icookie from the packet into the state object twice.