* Make note in comment about host/network order of msgid_t
The IKEv2 code was doing it badly, mixing up what was host and what
was network order. We will change this type from rt_raw so that the
packet marshalling code in out_struct() and in_struct() will deal
with it properly for us. For now we tried to reduce the ntohl/htonl
calls by making it host order for ikev2