* Backwards compatibly with broken SHA2 truncation using sha2_truncbug=yes
Instead of passing a fake algorithm all over the negotiation, track
this in the connection object, and only pass the fake algorithm in
the call to kernel_ops->add_sa(). This saves us from needing to hack
the out_attr() proposals and re-interpreting incoming proposals.

Currently, only the NETKEY stacks supports this.