* SHA2 default hash truncation fix and compat mode via sha2_truncbug=yes option
This uses the new struct xfrm_algo_auth instead of xfrm_algo to properly set
a 128bit hash truncation for sha2_256 in the kernel. Previously, the kernel
mistakenly used 96 bits. For compatibility, the old kernel behaviour can be
enabled using sha2_truncbug=yes in a connection configuration.

This patch should NOT be used on kernels that do not have the new XFRM code
backported, so ensure proper openswan package and kernel package dependancies.