ipsec-tools 0.8.0 mistakenly sets some NAT-OA fields that are defined in RFC1374 as "always zero". We define these as "ft_mbz" (Must Be Zero)
Introduce ft_zig value and implement android workaround with ft_zig

ft_zig (should be zero but ignore) is like ft_mbz (must be zero) except
we log a message and continue instead of aborting.

Currently, this is always logged. If we find this happens too often,
we can change the logging level to something non-default.

The RFC's state we should "ignore and continue" and not "abort"
when we receive a non-zero value for a field that "must be zero".