* Block rules created by openswan remain even after tunnel establishment
bugfix for #1334

Detail analysis

Problem text refers to block policy rules (bare shunts) which are created
by openswan and remain in kernel for ever until service is restarted.
This is happening if the user manualy adds policy rules.

If a manual policy rule is added, this will trigger an ACQUIRE
message. ACQUIRE will be handled by openswan by the following manner:

A valid conn will first be searched at initiate_ondemand_body function
at find_connection_for_clients.  If a valid connection is not found
then cannot_oppo function is call.  This function at the #ifdef KLIPS
part of code will call the replace_bare_shunt function. This function
is called with failure_shunt = FALSE and the transport protocol != 0.
What it is going on now is wrong!

openswan goes inside this function in order to delete the policy rule
which triggered the ACQUIRE msg (as it is stated from trace message). But
transport protocol is != 0. So instead of delete performs an addition!
By that way a bare shut or block policy rule is created in NETKEY stack
which cannot anymore be deleted by openswan.

In my opinion pluto, even if the user does not "properly uses" openswan
by adding a manual policy, should not add and moreover let those block
policy rules in the kernel.  Since those are added by openswan should also
properly delete them.  In NETKEY the default value of the parameter level
is "required" which means that no unecrypted traffic will go out until
an SA is fully established so a block rule is not required!  KLIPS might
need that but this part of code should not be executed under NETKEY stack.

Futhermore even in a proper use the code has a bug.  Lets assume now
that a very very simple host to host tunnel is erouted.

Pluto will add a policy (trap) to the kernel. When a ping is send from
one host to the other this will trigger an ACQUIRE message to tell
pluto to establish a tunnel.  Now the connection is found (proper use of
openswan) at initiate_ondemand_body function.  Code will go to assign_hold
function at line 860. assign_hold at the if statement eclipsable(sr)
will return true since it is a host to host tunnel and this later will
call free_bare_shunt function which will print "delete bare shunt:
null pointer".

In conclussion analysis is that those 2 IFDEF KLIPS code flows should
not be executed if NETKEY stack is used.  Below are traces from ACQUIRE
message in host to host and subnet to subnet erouted tunnel in "normal
use".

Conflicts:
	programs/pluto/initiate.c