Additional safety checks to addrtot(), inet_addrtot() and sin_addrtot()
These functions when used properly are bassed a char of size ADDRTOT_BUF
which is more then enough to strncat "<invalid>" into.  But scanning
tools don't know about these and show red flags in case something smaller
is passed.

So now we check if the dstlen passed in is smaller then sizeof("invalid")

At least if someone would call these functions with chars that are too
small, we just truncate the text "<invalid>".