* updown: Delete the source ip addres on down only for Cisco peer
We know we obtained (and need to lose) an IP when connecting using
a Cisco peer. There are other cases where an obtained sourceip should
not be removed (eg when moving from LAN to Wifi, using a setup where
you get the same sourceip, a case that Tuomo has)

This should cover most cases. I confirmed it works using the Red Hat
VPN.