wo#7572 .  ISAKMP_SA_established() should never release the current connection
ISAKMP_SA_established() tries to kill any connections that have the same IDs
on different end points.  The end points were compared by using sameaddr(),
which calls to addrcmp().  In NAT-T connections it is possible for the endpoint
to be undefined.

Due to a bug in addrcmp(), fixed in previous commit, it was possible for
ISAKMP_SA_established() to release the connection that was just established.

This commit adds an explicit test to skip the established connection from
removal consideration.