wo#7613 - transport conn w/ subnet marked as INVALID_CONFIG
In add_connection() we now call connection_has_valid_config() to test
that the connection is configured in a valid way.  If it is not, then
the c->policy gets a POLICY_INVALID_CONFIG bit set.

connection_has_valid_config() was added to perform checks on the whack
message.  Currently only one test is made: TRANSPORT mode conn cannot have
any subnets defined.