wo#5579 - encrypted notification processor mechanism
When an encrypted message arrives, it will match one of the states in
the v2_state_microcode_table[].  We do not know if we can handle it
until after it has been decrypted.  Once decrypted we may find that
it contains a notification from the peer indicating a failure.

This commit adds an ntf_processor callback to state_v2_microcode, which
is invoked post decrypt, by ikev2_process_encrypted_payloads(), if three
conditions are met:
(1) the payloads do not match req_enc_payloads,
(2) there was an encrypted v2N payload,
(3) there is a ntf_processor defined on the state.