wo#7347 - properly handle peer rejecting our DH group proposal
When there are multiple proposals with varied DH groups, we pick one and
create a matching v2KE payload.  However, if the peer refuses that DH
group, we need to retry the next proposal's DH group.

In this commit, the following changes are made to make this work:

(1) initiating a new conn, ikev2parent_outI1_withstate() will use
    guess_dhgroup() to get the next possible DH group to try.  The
    conn structure tracks of c->proposal_index and c->proposal_can_retry
    to later decide if a retry is possible, and to use the next
    available proposal's DH group.

(2) receiving a v2N_INVALID_KE_PAYLOAD notification for a SA_INIT
    message, will trigger a potential retry if more DH groups are
    available.  In this case, the c->proposal_index is incremented
    and the initiation process is restarted.  guess_dhgroup() will
    select the next DH group.