wo#7347 - validate group in accept_v2_KE()
Refactor accept_v2_KE() to do additional validation.

In this function, 'v2ke->isak_group' is the DH group the remote has
selected and is the group used for the v2KE payload.  We may have
selected a different DH group from the proposal.

This change will detect such a misunderstanding and send the peer a
v2N_INVALID_KE_PAYLOAD to let them know we didn't like the proposal.