wo#8784 - IKEv2 will not add padding to INIT exchange
we used to call close_message() as part of I1/R1 out messages, which
added padding if the payloadds were not all multiples of 4 octets.  This
became a problem after we started sending out CERTREQ, which is never a
multiple of 4 octets.

This change has been validated to interoperate with strongSwan 5.7.2 and
OpenSWAN 2.6.51.