Fix -Wstringop-truncation errors
In most cases, the strings are treated as null terminated elsewhere so
limiting to sizeof()-1 is correct.

In the case of req.ifr_name, the kernel adds null termination if the
name isn't already.

sun_path doesn't need to be null terminated, so this does limit the path
to one less than maximum.